Restructured component files
Following convention of putting components inside a ‘components’ directory.
