+++ packages/lib-classifier/src/components/Classifier/components/SubjectViewer/components/VideoController/index.js
@@ -1,3 +1,1 @@
 export { default } from './VideoController'
-export { default } from './FormattedTime'
-export { default } from './Slider'
